fixes to more ogonecks, anchors, sharp/flat point fixes, diacritics
I also found that Tdiagonalstroke and Tdiagonalstroke.sc was suffering from an optical issusion of the type seen in Lslash. So I decomposed it in roman and iatalic and corrected it. Marc pointed out an inconsistemcy in the construction of Bold master 144pt N and I made it more consistent with the light 144pt with splayed out stems that get a bit heavier at the serifs.